Partial fix for bug #700448.
NULL is returned from o_*_read and o_read_buffer functions if parsing of the input buffer fails.
An error object is returned in these cases, describing in detail what error occured.

gnetlist is updated to exit with status code 2 if the input file can't be read.

gschem is updated to report errors when using o_read_buffer.

TODO:
* Prone to segfaults on bad input files (separate patch)